% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_cox_function.R
\name{bootstrap_cox_function}
\alias{bootstrap_cox_function}
\title{The bootstrap_cox function}
\usage{
bootstrap_cox_function(variables, number, model_data, ndraws, l = "lambda.1se")
}
\description{
The bootstrap_cox function
}
