% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_log_function.R
\name{bootstrap_log_function}
\alias{bootstrap_log_function}
\title{The bootstrap_log function}
\usage{
bootstrap_log_function(variables, number, model_data, ndraws, l = "lambda.1se")
}
\description{
The bootstrap_log function
}
