% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characteristics_function.R
\name{char}
\alias{char}
\title{The Characteristics function}
\usage{
char(dataframe, key, exclude)
}
\arguments{
\item{dataframe}{dataframe with the p-values of all variables (outDig)}

\item{key}{dataframe with two columns; 1: list of variable names with (current) working name. 2: list of variable names with (desired) output name.}

\item{exclude}{variables which to exclude from the summary statistics}
}
\description{
This function creates characteristics (mean(sd) or %) using the createtableone function, displaying all summary statistics of the variables of interest
Automatically saves the output as a csv file.
}
\examples{
characteristics <- char(dataGrouped, "key", "exclude"); plot1; plot1 + scale_y_continuous(expand = c(0,0), limits = c(0, 50)); ggsave("manhattan.png")
plotManhattan()
}
\keyword{Characteristics}
