% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationPlot_function.R
\name{correlating}
\alias{correlating}
\title{The correlationPlot function}
\usage{
correlating(variables, df, key, nColors)
}
\arguments{
\item{variables}{dataframe containing the siginificant variables}

\item{df}{dataframe (data)}

\item{key}{dataframe containing}

\item{nColors}{number of colors per direction (4 per direction equals 8 colors in total)}
}
\description{
This function creates a correlation plot of the variables of interest, using partial Pearson correlations and adjusting for age and sex
The variables in the plot are ordered using a hierargical clustering method.
This function outputs a correlation plot and a correlation table. Both are automatically saved in the workspace.
}
\examples{
plot3 <-  correlating(outCon, dataOut, key, 4)
}
\keyword{clustering}
\keyword{correlation}
\keyword{hierargical}
\keyword{matrix,}
\keyword{plot,}
