% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dividing_function.R
\name{dividing}
\alias{dividing}
\title{The dividing function}
\usage{
dividing(split, marge, id, dataframe)
}
\arguments{
\item{split}{Variable on which to randomly split the dataframe. put in parenteces ("")}

\item{marge}{Amount of variation between groups (1 = equal groups). Smaller marge may lead to longer runtime}

\item{id}{Unique subject ID}

\item{dataframe}{Dataframe}
}
\description{
This function creates the variable group, which divides the dataframe into two different groups; Group A and Group B.
}
\examples{
dataGrouped <- dividing("zip", 1, "PSEUDOIDEXT", data)
dividing()
}
\keyword{divide,}
\keyword{group}
