% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharing_legend_function.R
\name{grid_arrange_shared_legend}
\alias{grid_arrange_shared_legend}
\title{The sharing legend function}
\usage{
grid_arrange_shared_legend(
  ...,
  nrow = length(list(...)),
  ncol = 1,
  position = c("bottom", "right")
)
}
\description{
The sharing legend function
}
