% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lassoing_function.R
\name{lassoing}
\alias{lassoing}
\title{The lassoing function}
\usage{
lassoing(
  variables,
  subgroup,
  dataframe,
  number = 1,
  l = "lambda.1se",
  name = subgroup
)
}
\arguments{
\item{variables}{A dataframe with a column for all variables of interest and their respective groups}

\item{subgroup}{The group which includes all variables of interest for that group}

\item{dataframe}{Dataframe}

\item{number}{Number of observation a variable needs in order to be included. Some variables with many missing may reduce the sample size of the total analysis and may be considered to exclude. Defaults to n=1.}

\item{l}{The way lambda should be picked (i.e. "lamda.min" or "lambda.1se")}

\item{name}{The name of the output file. Defaults to name of subgroup (warning: when including multiple subgroups, the default leads to error)}
}
\description{
This function performs lasso regression on the different pre-defined groups. 
All variables of the respective group are included, after which Lasso regression is performed to perform variable selection. 
A 10-fold cross-validation is used. The final model is printed as "name"Lasso.csv.
The output of this function can be used to subsequently plot the ROC curves.
}
\examples{
anthro <- lassoing(dfVulcanoTwo, "Anthropometrics", dataCon)
lassoing()
}
\keyword{group}
