% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mEff_function.R
\name{mEff}
\alias{mEff}
\title{The mEff function}
\usage{
mEff(key, variables, groupVars, df)
}
\arguments{
\item{key}{dataframe containing workName and group of every variable (key)}

\item{variables}{dataframe containing the siginificant variables (dataCon)}

\item{groupVars}{column within "key" containing groups of interest (e.g. "group2")}

\item{df}{dataframe (data)}
}
\description{
This function calculates the effective number of variables after taking into account correlation
}
\examples{
mEff(key, outCon, "group2", dataCon)
}
\keyword{effective}
\keyword{number}
\keyword{of}
\keyword{variables}
