% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeCox_function.R
\name{pipe_cox}
\alias{pipe_cox}
\title{The pipeCox function}
\usage{
pipe_cox(disease, exclude, group, dataframe)
}
\arguments{
\item{disease}{name coding for disease variable (e.g. "dmNewDig")}

\item{exclude}{Variables which are not of interest}

\item{group}{Variable which identifies respective group (created with grouping function)}

\item{dataframe}{dataframe (dataDig, created with grouping function)}
}
\description{
This function runs a cox-regression model pipeline for all variables of interest, adjusting for age and sex.
The pipeline starts with a cox-regression in group A, after which all variables with a False Discovery Rate (BY) of alpha < 0.05 are selected and the analysis is repeated in group B. Pathways go both ways (A --> B and B --> A)
When the variable is significant in none of the pathways, models == "none", when significant in the first but not second model, models == "one", when significant in both pathways models == "two".
Output is a dataframe with all variables tested, estimate and p-value from the whole dataset, number of complete cases and significance in pathways
}
\examples{
outDig <- pipe_cox("dmNewDig", exclude, group, dataDig)
pipe_cox()
}
\keyword{pipeline}
