% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeCoxContinuous_function.R
\name{pipe_cox_continuous}
\alias{pipe_cox_continuous}
\title{The pipe_cox_continuous function}
\usage{
pipe_cox_continuous(disease, outDigSig, dataframe, key)
}
\arguments{
\item{disease}{name coding for disease variable (e.g. "dmNewDig")}

\item{dataframe}{dataframe with continuous variables (dataCon, created with grouping function)}

\item{key}{key with workName and outName, for output of proper names}

\item{outDig}{output file from the digotomous pipeline (pipeCox function), optionally filtered for solely significant variables}
}
\description{
This function runs a cox-regression model pipeline for all variables of interest, adjusting for age and sex.
The pipeline starts with a cox-regression in group A, after which all variables with a False Discovery Rate (BY) of alpha < 0.05 are selected and the analysis is repeated in group B. Pathways go both ways (A --> B and B --> A)
When the variable is significant in none of the pathways, models == "none", when significant in the first but not second model, models == "one", when significant in both pathways models == "two".
Output is a dataframe with all variables tested, estimate and p-value from the whole dataset, number of complete cases and significance in pathways
This function uses the original factors to improve interpretability. Reference groups have to be set by hand!
}
\examples{
outDigSig <- outDig \%>\% filter(models != "none"); outCon <- pipe_cox_continuous(outDigSig, dataCon)
pipe_cox_continuous()
}
\keyword{pipeline}
