% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeLogContinuous_function.R
\name{pipe_log_continuous}
\alias{pipe_log_continuous}
\title{The pipe_log function for continuous traits}
\usage{
pipe_log_continuous(disease, outDigSig, dataframe, key)
}
\description{
The pipe_log function for continuous traits
}
