% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotManhattan_function.R
\name{plotManhattan}
\alias{plotManhattan}
\title{The plotManhattan function}
\usage{
plotManhattan(dataframe, gr, significant, p)
}
\arguments{
\item{dataframe}{dataframe with the p-values of all variables (outDig)}

\item{gr}{variable which holds info in which group the respective variable is}

\item{significant}{variable which holds info in how many pathways a variable was significant (one, two, none; models variable)}

\item{p}{variable which holds the p-value of the variables (pValue)}
}
\description{
This function creates a Manhattan plot, displaying all p-values of the variables of interest
This function requires the individual variables to be classified into a grouping variable
}
\examples{
plot1 <- plotManhattan(outDig, "group", "models", "pValue"); plot1; plot1 + scale_y_continuous(expand = c(0,0), limits = c(0, 50)); ggsave("manhattan.png")
plotManhattan()
}
\keyword{Manhattan}
