% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVulcano_function.R
\name{plotVulcano}
\alias{plotVulcano}
\title{The plotVulcano function}
\usage{
plotVulcano(dataframe, est, p, gr)
}
\arguments{
\item{dataframe}{dataframe with the p-values of all variables (outDig)}

\item{est}{variable which holds the estimates of the variables (estimate)}

\item{p}{variable which holds the p-value of the variables (pValue)}

\item{gr}{variable which holds info in which group the respective variable is}
}
\description{
This function creates a Vulcano plot, displaying all estimates and p-values of the variables of interest
This function requires the individual variables to be classified into a grouping variable
}
\examples{
plot2 <- plotVulcano(dfVulcano, "estimate", "pValue", "group2")
}
\keyword{Vulcano}
