% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_function.R
\name{to_con}
\alias{to_con}
\title{The dummy function}
\usage{
to_con(dataframe, exclude)
}
\arguments{
\item{dataframe}{Dataframe which is to be used as input.}

\item{exclude}{Variables which do not need to be converted.}
}
\description{
This document describes to functions. The first function, to_con divides the variables of interest in factors and continuous variables.
All variables with less than 20 unique values are categorized as factor. Numerical variables are recentered and scaled using standard deviation.
The second function, to_dig succcessively creates dummy variables of all factors.
}
\examples{
exeptions <- c("PSEUDOIDEXT", "idFU", "group", "dmNew", "dmNewDig", "time"); dataCon <- to_con(dataGrouped, exeptions); dataDig <- to_dig(dataGrouped, exeptions)
to_con(), to_dig()
}
\keyword{dummie}
\keyword{dummy,}
